В довесок, перенес логику добавления клиента в репозиторий клиентов
