+++ app/src/main/java/ru/nsu/fit/di/RepositoryModule.kt
@@ -3,7 +3,6 @@ package ru.nsu.fit.di
 import dagger.Binds
 import dagger.Module
 import ru.nsu.fit.data.repository.*
-import ru.nsu.fit.domain.model.Customer
 import ru.nsu.fit.domain.repository.*
 
 @Module
@@ -24,7 +23,7 @@ interface RepositoryModule {
     fun bindWheelSizeRepository(repo: WheelSizeRepositoryImpl): WheelSizeRepository
 
     @Binds
-    fun bindSalesRepository(repo: SalesRepositoryImpl): SalesRepository
+    fun bindSalesRepository(repo: SalesManagerImpl): SalesRepository
 
     @Binds
     fun bindCustomerRepository(repo: CustomerRepositoryImpl): CustomerRepository
